修复g++上的编译警告

+++ wedge/binance/binance_http_client.cc
@@ -82,7 +82,7 @@ class URLBuilder {
  private:
   URLBuilder& set(std::string_view key, int64_t value) {
     char buffer[128];
-    sprintf(buffer, "%lld", value);
+    sprintf(buffer, "%lld", (long long) value);
     return set(key, buffer);
   }
